Make the fidelity gate observe Lc0 and the served dtype
The first gate run parsed no priors: quit reached Lc0 before its search printed
anything, so the driver now waits for bestmove. Lc0 reports priors after its own
policy softmax temperature, so the gate pins it to 1 to compare raw network
output. It also checks that the traced wrapper gives the same answer batched as
alone, since tracing an ONNX conversion can bake the sample batch size into its
reshapes, and reports bfloat16 against float32 agreement because evaluation
serves the network in bfloat16. The wrapper no longer casts its input, so it
runs in whatever dtype the pipeline loads it in.

Validated: test_lc0_teacher_tooling 20 passed; ruff clean.

+++ py/tools/distill_build_dataset.py
@@ -267,7 +267,7 @@ def generate_records(
         )
         if isinstance(teacher, Lc0Teacher):
             # The teacher reads Lc0's planes; the record keeps the project's, which the student trains on.
-            decoded = decode_lc0_planes(tuple(slot.position.lc0_packed_encoding() for slot in slots))
+            decoded = decode_lc0_planes(tuple(slot.position.lc0_packed_encoding() for slot in slots)).astype(np.float32)
         else:
             decoded = decode_packed_inputs(CHESS_STATE_CONTRACT, packed_states)
         with torch.inference_mode():
